requirements.txt
Pillow>=10.0.0
tqdm>=4.66.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
pydocstyle>=6.3.0
pandas>=2.0.0

